+++ src/components/Layout/Layout.js
@@ -6,9 +6,9 @@ import Button from '@material-ui/core/Button';
 import Header from './Header';
 import Footer from './Footer';
 import Prayers from '../Prayers';
-import {usePrayer} from '../../hooks/api-hooks';
+import {usePrayer, useDrawer} from '../../hooks/api-hooks';
 import {UserContext} from '../../store/context/userContext';
-
+import Menus from '../Menus'
 
 
 const useStyles = makeStyles(theme => ({
@@ -22,6 +22,7 @@ const useStyles = makeStyles(theme => ({
   }));
 
 const Layout = ({country, region, place, pdate, startup}) => {
+    const [drawerOpen, handleDrawerToggle] = useDrawer();
     const {setTz} = useContext(UserContext);
     const [data] = usePrayer({region: region, country: country, place : place, date : pdate});
     const {timezone} = (data && data.data && data.data.meta)?data.data.meta:'Europe/AmsterDAM';
@@ -30,10 +31,12 @@ const Layout = ({country, region, place, pdate, startup}) => {
     if (data && data.data && data.data.meta && data.code === 200) {
         return (
             <>
+                <Menus drawerOpen={drawerOpen} handleDrawerToggle={handleDrawerToggle}/>
                 <Header 
                     startup={startup} 
                     place={localStorage.getItem(`padachone:place`)}
                     pdate={data.data.date.readable}
+                    handleDrawerToggle={handleDrawerToggle}
                 />
                <Prayers prdata={data} />