+++ src/components/Travel/Travel.js
@@ -30,7 +30,8 @@ const Travel = (props) => {
                             />
                         </>
                         :<div>Getting the location data&hellip; </div>
-            }           
+            }  
+                     {/* me */}
             {/* <Traveltimes lat="52.31406610552598" lon="4.946411339519716" music={music} setMusic={setMusic} volume={volume} setVolume={setVolume}/> */}
         </>
     )