+++ src/utils/constants.js
@@ -2,3 +2,16 @@ export const P_EMAIL = "padachone_email";
 export const BING_API = "ArNqsYDx-rtxpMbR4ddz8SyY4-dv8-JK35KErFW3GIU7_UwgaCVz8Bj4iKy4X-JP";
 export const PRAYERS_ARR = ['Fajr', 'Dhuhr', 'Asr', 'Maghrib', 'Isha'];
 export const REPORT_FB = "Report Crash";
+export const P_MENUS = [
+    {label : 'Home', page : 'Home'}, 
+    {label : 'Travel Onboard', page : 'Travel'},
+    {label : 'Subscribe', page: 'setmodal'},    
+    // {label : 'Fine Tune', page : 'Fine Tune'}, 
+    // {label : 'Travel Times', page : 'Travel'}, 
+    // {label : 'Nearest Mosque', page : 'Nearest Mosque'}, 
+    // {label : 'Map', page : 'Map'} 
+];
+export const P_MENUS_SEC = [
+    {label : 'Messy Lab', page: 'Lab'},
+    {label : 'Start Over', page: 'callfunc'}  
+];